reader is done

+++ includes/lem_in.h
@@ -6,7 +6,7 @@
 /*   By: dtimeon <dtimeon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/19 15:02:48 by fsinged           #+#    #+#             */
-/*   Updated: 2019/10/21 13:27:54 by fsinged          ###   ########.fr       */
+/*   Updated: 2019/10/21 14:38:38 by fsinged          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ t_options			*read_options(int argc, char **argv);
 int					validation(char **data, char ***rooms, char ***ways);
 void				ft_error(char *str);
 
-t_list				*ft_newlist(char **content, int content_size);
+t_list				*ft_newlist(char **content, size_t content_size);
 
 /*
 ** fill.c
